almost done with jdbc exercise

+++ src/main/java/ListAdsDao.java
@@ -11,7 +11,7 @@ public class ListAdsDao implements Ads {
         return ads;
     }
 
-    public Long insert(Ad ad) {
+    public int insert(Ad ad) {
         // make sure we have ads
         if (ads == null) {
             ads = generateAds();
